+++ connect/models/schemas.py
@@ -18,7 +18,7 @@ class BaseSchema(Schema):
 
 class ActivationSchema(BaseSchema):
     link = fields.Str(allow_none=True)
-    message = fields.Str()
+    message = fields.Str(allow_none=True)
     date = fields.DateTime(allow_none=True)
 
     @post_load
@@ -29,7 +29,7 @@ class ActivationSchema(BaseSchema):
 
 class AgreementStatsSchema(BaseSchema):
     contracts = fields.Int(allow_none=True)
-    versions = fields.Int()
+    versions = fields.Int(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -38,7 +38,7 @@ class AgreementStatsSchema(BaseSchema):
 
 
 class CompanySchema(BaseSchema):
-    name = fields.Str()
+    name = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -59,10 +59,10 @@ class PhoneNumberSchema(BaseSchema):
 
 
 class ContactSchema(BaseSchema):
-    email = fields.Str()
+    email = fields.Str(allow_none=True)
     first_name = fields.Str(allow_none=True)
     last_name = fields.Str(allow_none=True)
-    phone_number = fields.Nested(PhoneNumberSchema)
+    phone_number = fields.Nested(PhoneNumberSchema, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -71,12 +71,12 @@ class ContactSchema(BaseSchema):
 
 
 class ContactInfoSchema(BaseSchema):
-    address_line1 = fields.Str()
+    address_line1 = fields.Str(allow_none=True)
     address_line2 = fields.Str(allow_none=True)
-    city = fields.Str()
-    contact = fields.Nested(ContactSchema)
-    country = fields.Str()
-    postal_code = fields.Str()
+    city = fields.Str(allow_none=True)
+    contact = fields.Nested(ContactSchema, allow_none=True)
+    country = fields.Str(allow_none=True)
+    postal_code = fields.Str(allow_none=True)
     state = fields.Str(allow_none=True)
 
     @post_load
@@ -86,8 +86,8 @@ class ContactInfoSchema(BaseSchema):
 
 
 class ValueChoiceSchema(Schema):
-    value = fields.Str()
-    label = fields.Str()
+    value = fields.Str(allow_none=True)
+    label = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -96,10 +96,10 @@ class ValueChoiceSchema(Schema):
 
 
 class ConstraintsSchema(BaseSchema):
-    hidden = fields.Bool()
-    required = fields.Bool()
-    choices = fields.Nested(ValueChoiceSchema, many=True)
-    unique = fields.Bool()
+    hidden = fields.Bool(allow_none=True)
+    required = fields.Bool(allow_none=True)
+    choices = fields.Nested(ValueChoiceSchema, many=True, allow_none=True)
+    unique = fields.Bool(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -108,8 +108,8 @@ class ConstraintsSchema(BaseSchema):
 
 
 class DocumentSchema(BaseSchema):
-    title = fields.Str()
-    url = fields.Str()
+    title = fields.Str(allow_none=True)
+    url = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -118,9 +118,9 @@ class DocumentSchema(BaseSchema):
 
 
 class DownloadLinkSchema(BaseSchema):
-    title = fields.Str()
-    url = fields.Str()
-    visible_for = fields.Str()
+    title = fields.Str(allow_none=True)
+    url = fields.Str(allow_none=True)
+    visible_for = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -129,7 +129,7 @@ class DownloadLinkSchema(BaseSchema):
 
 
 class UserSchema(BaseSchema):
-    name = fields.Str()
+    name = fields.Str(allow_none=True)
     email = fields.Str(allow_none=True)
 
     @post_load
@@ -149,17 +149,17 @@ class EventSchema(BaseSchema):
 
 
 class EventsSchema(BaseSchema):
-    created = fields.Nested(EventSchema)
-    inquired = fields.Nested(EventSchema)
-    pended = fields.Nested(EventSchema)
-    validated = fields.Nested(EventSchema)
-    updated = fields.Nested(EventSchema)
-    approved = fields.Nested(EventSchema)
-    uploaded = fields.Nested(EventSchema)
-    submitted = fields.Nested(EventSchema)
-    accepted = fields.Nested(EventSchema)
-    rejected = fields.Nested(EventSchema)
-    closed = fields.Nested(EventSchema)
+    created = fields.Nested(EventSchema, allow_none=True)
+    inquired = fields.Nested(EventSchema, allow_none=True)
+    pended = fields.Nested(EventSchema, allow_none=True)
+    validated = fields.Nested(EventSchema, allow_none=True)
+    updated = fields.Nested(EventSchema, allow_none=True)
+    approved = fields.Nested(EventSchema, allow_none=True)
+    uploaded = fields.Nested(EventSchema, allow_none=True)
+    submitted = fields.Nested(EventSchema, allow_none=True)
+    accepted = fields.Nested(EventSchema, allow_none=True)
+    rejected = fields.Nested(EventSchema, allow_none=True)
+    closed = fields.Nested(EventSchema, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -168,7 +168,7 @@ class EventsSchema(BaseSchema):
 
 
 class HubInstanceSchema(BaseSchema):
-    type = fields.Str()
+    type = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -177,8 +177,8 @@ class HubInstanceSchema(BaseSchema):
 
 
 class HubStatsSchema(BaseSchema):
-    connections = fields.Int()
-    marketplaces = fields.Int()
+    connections = fields.Int(allow_none=True)
+    marketplaces = fields.Int(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -187,12 +187,12 @@ class HubStatsSchema(BaseSchema):
 
 
 class HubSchema(BaseSchema):
-    name = fields.Str()
-    company = fields.Nested(CompanySchema)
+    name = fields.Str(allow_none=True)
+    company = fields.Nested(CompanySchema, allow_none=True)
     description = fields.Str(allow_none=True)
-    instance = fields.Nested(HubInstanceSchema)
-    events = fields.Nested(EventsSchema)
-    stats = fields.Nested(HubStatsSchema)
+    instance = fields.Nested(HubInstanceSchema, allow_none=True)
+    events = fields.Nested(EventsSchema, allow_none=True)
+    stats = fields.Nested(HubStatsSchema, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -211,8 +211,8 @@ class ExternalIdField(fields.Field):
 
 
 class ExtIdHubSchema(Schema):
-    hub = fields.Nested(HubSchema, only=('id', 'name'))
-    external_id = ExternalIdField()
+    hub = fields.Nested(HubSchema, only=('id', 'name'), allow_none=True)
+    external_id = ExternalIdField(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -221,10 +221,10 @@ class ExtIdHubSchema(Schema):
 
 
 class RenewalSchema(BaseSchema):
-    from_ = fields.DateTime(load_from='from')
-    to = fields.DateTime()
-    period_delta = fields.Int()
-    period_uom = fields.Str()
+    from_ = fields.DateTime(load_from='from', allow_none=True)
+    to = fields.DateTime(allow_none=True)
+    period_delta = fields.Int(allow_none=True)
+    period_uom = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -253,13 +253,13 @@ class QuantityField(fields.Field):
 
 
 class MarketplaceSchema(BaseSchema):
-    name = fields.Str()
-    description = fields.Str()
-    active_contracts = fields.Int()
-    icon = fields.Str()
-    owner = fields.Nested(CompanySchema, only=('id', 'name'))
-    hubs = fields.Nested(ExtIdHubSchema, many=True)
-    zone = fields.Str()
+    name = fields.Str(allow_none=True)
+    description = fields.Str(allow_none=True)
+    active_contracts = fields.Int(allow_none=True)
+    icon = fields.Str(allow_none=True)
+    owner = fields.Nested(CompanySchema, only=('id', 'name'), allow_none=True)
+    hubs = fields.Nested(ExtIdHubSchema, many=True, allow_none=True)
+    zone = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -268,9 +268,9 @@ class MarketplaceSchema(BaseSchema):
 
 
 class CountrySchema(BaseSchema):
-    name = fields.Str()
-    icon = fields.Str()
-    zone = fields.Str()
+    name = fields.Str(allow_none=True)
+    icon = fields.Str(allow_none=True)
+    zone = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -279,9 +279,9 @@ class CountrySchema(BaseSchema):
 
 
 class ParamSchema(BaseSchema):
-    name = fields.Str()
-    description = fields.Str()
-    type = fields.Str()
+    name = fields.Str(allow_none=True)
+    description = fields.Str(allow_none=True)
+    type = fields.Str(allow_none=True)
     value = fields.Str(allow_none=True)
     value_error = fields.Str(allow_none=True)
     value_choice = fields.Str(many=True, allow_none=True)
@@ -303,8 +303,8 @@ class ParamSchema(BaseSchema):
 
 
 class ItemSchema(BaseSchema):
-    mpn = fields.Str()
-    quantity = QuantityField()
+    mpn = fields.Str(allow_none=True)
+    quantity = QuantityField(allow_none=True)
     old_quantity = QuantityField(allow_none=True)
     renewal = fields.Nested(RenewalSchema, allow_none=True)
     params = fields.Nested(ParamSchema, many=True, allow_none=True)
@@ -322,20 +322,20 @@ class ItemSchema(BaseSchema):
 
 
 class AgreementSchema(BaseSchema):
-    type = fields.Str()
-    title = fields.Str()
-    description = fields.Str()
-    created = fields.DateTime()
-    updated = fields.DateTime()
-    owner = fields.Nested(CompanySchema)
+    type = fields.Str(allow_none=True)
+    title = fields.Str(allow_none=True)
+    description = fields.Str(allow_none=True)
+    created = fields.DateTime(allow_none=True)
+    updated = fields.DateTime(allow_none=True)
+    owner = fields.Nested(CompanySchema, allow_none=True)
     stats = fields.Nested(AgreementStatsSchema, allow_none=True)
     author = fields.Nested(UserSchema, allow_none=True)
-    version = fields.Int()
-    active = fields.Bool()
-    link = fields.Str()
-    version_created = fields.DateTime()
-    version_contracts = fields.Int()
-    agreements = fields.Nested('AgreementSchema', many=True)
+    version = fields.Int(allow_none=True)
+    active = fields.Bool(allow_none=True)
+    link = fields.Str(allow_none=True)
+    version_created = fields.DateTime(allow_none=True)
+    version_contracts = fields.Int(allow_none=True)
+    agreements = fields.Nested('AgreementSchema', many=True, allow_none=True)
     parent = fields.Nested('AgreementSchema', only=('id', 'name'), allow_none=True)
     marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'), allow_none=True)
     name = fields.Str(allow_none=True)
@@ -347,19 +347,19 @@ class AgreementSchema(BaseSchema):
 
 
 class ContractSchema(BaseSchema):
-    name = fields.Str()
-    version = fields.Int()
-    type = fields.Str()
-    status = fields.Str()
-    agreement = fields.Nested(AgreementSchema, only=('id', 'name'))
+    name = fields.Str(allow_none=True)
+    version = fields.Int(allow_none=True)
+    type = fields.Str(allow_none=True)
+    status = fields.Str(allow_none=True)
+    agreement = fields.Nested(AgreementSchema, only=('id', 'name'), allow_none=True)
     marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'), allow_none=True)
     owner = fields.Nested(CompanySchema, only=('id', 'name'), allow_none=True)
-    creator = fields.Nested(UserSchema, only=('id', 'name'))
-    created = fields.DateTime()
-    updated = fields.DateTime()
+    creator = fields.Nested(UserSchema, only=('id', 'name'), allow_none=True)
+    created = fields.DateTime(allow_none=True)
+    updated = fields.DateTime(allow_none=True)
     enrolled = fields.DateTime(allow_none=True)
-    version_created = fields.DateTime()
-    activation = fields.Nested(ActivationSchema)
+    version_created = fields.DateTime(allow_none=True)
+    activation = fields.Nested(ActivationSchema, allow_none=True)
     signee = fields.Nested(UserSchema, only=('id', 'name'), allow_none=True)
 
     @post_load
@@ -369,8 +369,8 @@ class ContractSchema(BaseSchema):
 
 
 class ProductConfigurationSchema(BaseSchema):
-    suspend_resume_supported = fields.Bool()
-    requires_reseller_information = fields.Bool()
+    suspend_resume_supported = fields.Bool(allow_none=True)
+    requires_reseller_information = fields.Bool(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -379,10 +379,10 @@ class ProductConfigurationSchema(BaseSchema):
 
 
 class CustomerUiSettingsSchema(BaseSchema):
-    description = fields.Str()
-    getting_started = fields.Str()
-    download_links = fields.Nested(DownloadLinkSchema, many=True)
-    documents = fields.Nested(DocumentSchema, many=True)
+    description = fields.Str(allow_none=True)
+    getting_started = fields.Str(allow_none=True)
+    download_links = fields.Nested(DownloadLinkSchema, many=True, allow_none=True)
+    documents = fields.Nested(DocumentSchema, many=True, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -391,7 +391,7 @@ class CustomerUiSettingsSchema(BaseSchema):
 
 
 class ProductFamilySchema(BaseSchema):
-    name = fields.Str()
+    name = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -400,7 +400,7 @@ class ProductFamilySchema(BaseSchema):
 
 
 class ProductCategorySchema(BaseSchema):
-    name = fields.Str()
+    name = fields.Str(allow_none=True)
     parent = fields.Nested('ProductCategorySchema', allow_none=True)
     children = fields.Nested('ProductCategorySchema', many=True, allow_none=True)
     family = fields.Nested(ProductFamilySchema, allow_none=True)
@@ -412,8 +412,8 @@ class ProductCategorySchema(BaseSchema):
 
 
 class ProductStatsInfoSchema(BaseSchema):
-    distribution = fields.Int()
-    sourcing = fields.Int()
+    distribution = fields.Int(allow_none=True)
+    sourcing = fields.Int(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -422,9 +422,9 @@ class ProductStatsInfoSchema(BaseSchema):
 
 
 class ProductStatsSchema(BaseSchema):
-    listing = fields.Int()
-    agreements = fields.Nested(ProductStatsInfoSchema)
-    contracts = fields.Nested(ProductStatsInfoSchema)
+    listing = fields.Int(allow_none=True)
+    agreements = fields.Nested(ProductStatsInfoSchema, allow_none=True)
+    contracts = fields.Nested(ProductStatsInfoSchema, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -434,10 +434,10 @@ class ProductStatsSchema(BaseSchema):
 
 class ProductConfigurationParameterSchema(BaseSchema):
     value = fields.Str(allow_none=True)
-    parameter = fields.Nested(ParamSchema)
+    parameter = fields.Nested(ParamSchema, allow_none=True)
     marketplace = fields.Nested(MarketplaceSchema, allow_none=True)
     item = fields.Nested(ItemSchema, allow_none=True)
-    events = fields.Nested(EventsSchema)
+    events = fields.Nested(EventsSchema, allow_none=True)
     constraints = fields.Nested(ConstraintsSchema, allow_none=True)
 
     @post_load
@@ -447,14 +447,14 @@ class ProductConfigurationParameterSchema(BaseSchema):
 
 
 class ProductSchema(BaseSchema):
-    name = fields.Str()
-    icon = fields.Str()
-    short_description = fields.Str()
-    detailed_description = fields.Str()
-    version = fields.Int()
+    name = fields.Str(allow_none=True)
+    icon = fields.Str(allow_none=True)
+    short_description = fields.Str(allow_none=True)
+    detailed_description = fields.Str(allow_none=True)
+    version = fields.Int(allow_none=True)
     published_at = fields.DateTime(allow_none=True)
-    configurations = fields.Nested(ProductConfigurationSchema)
-    customer_ui_settings = fields.Nested(CustomerUiSettingsSchema)
+    configurations = fields.Nested(ProductConfigurationSchema, allow_none=True)
+    customer_ui_settings = fields.Nested(CustomerUiSettingsSchema, allow_none=True)
     category = fields.Nested(ProductCategorySchema, allow_none=True)
     owner = fields.Nested(CompanySchema, allow_none=True)
     latest = fields.Bool(allow_none=True)
@@ -467,9 +467,9 @@ class ProductSchema(BaseSchema):
 
 
 class ServerErrorResponseSchema(Schema):
-    error_code = fields.Str()
+    error_code = fields.Str(allow_none=True)
     params = fields.Dict(allow_none=True)
-    errors = fields.List(fields.Str())
+    errors = fields.Str(many=True, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -478,9 +478,9 @@ class ServerErrorResponseSchema(Schema):
 
 
 class TemplateSchema(BaseSchema):
-    name = fields.Str()
-    representation = fields.Str()
-    body = fields.Str()
+    name = fields.Str(allow_none=True)
+    representation = fields.Str(allow_none=True)
+    body = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -489,7 +489,7 @@ class TemplateSchema(BaseSchema):
 
 
 class ConfigurationSchema(BaseSchema):
-    params = fields.Nested(ParamSchema, many=True)
+    params = fields.Nested(ParamSchema, many=True, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -498,10 +498,10 @@ class ConfigurationSchema(BaseSchema):
 
 
 class TierAccountSchema(BaseSchema):
-    name = fields.Str()
-    contact_info = fields.Nested(ContactInfoSchema)
-    external_id = ExternalIdField()
-    external_uid = fields.Str()
+    name = fields.Str(allow_none=True)
+    contact_info = fields.Nested(ContactInfoSchema, allow_none=True)
+    external_id = ExternalIdField(allow_none=True)
+    external_uid = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -510,7 +510,7 @@ class TierAccountSchema(BaseSchema):
 
 
 class TierAccountsSchema(Schema):
-    customer = fields.Nested(TierAccountSchema)
+    customer = fields.Nested(TierAccountSchema, allow_none=True)
     tier1 = fields.Nested(TierAccountSchema, allow_none=True)
     tier2 = fields.Nested(TierAccountSchema, allow_none=True)
 
@@ -521,11 +521,11 @@ class TierAccountsSchema(Schema):
 
 
 class ConnectionSchema(BaseSchema):
-    type = fields.Str()
-    provider = fields.Nested(CompanySchema, only=('id', 'name'))
-    vendor = fields.Nested(CompanySchema, only=('id', 'name'))
-    product = fields.Nested(ProductSchema)
-    hub = fields.Nested(HubSchema)
+    type = fields.Str(allow_none=True)
+    provider = fields.Nested(CompanySchema, only=('id', 'name'), allow_none=True)
+    vendor = fields.Nested(CompanySchema, only=('id', 'name'), allow_none=True)
+    product = fields.Nested(ProductSchema, allow_none=True)
+    hub = fields.Nested(HubSchema, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -534,20 +534,20 @@ class ConnectionSchema(BaseSchema):
 
 
 class AssetSchema(BaseSchema):
-    status = fields.Str()
-    external_id = ExternalIdField()
+    status = fields.Str(allow_none=True)
+    external_id = ExternalIdField(allow_none=True)
     events = fields.Nested(EventsSchema, allow_none=True)
     external_uid = fields.Str(allow_none=True)
     external_name = fields.Str(allow_none=True)
-    product = fields.Nested(ProductSchema, only=('id', 'name'))
+    product = fields.Nested(ProductSchema, only=('id', 'name'), allow_none=True)
     connection = fields.Nested(
-        ConnectionSchema, only=('id', 'type', 'provider', 'vendor'),
+        ConnectionSchema, only=('id', 'type', 'provider', 'vendor'), allow_none=True
     )
     contract = fields.Nested(ContractSchema, allow_none=True)
     marketplace = fields.Nested(MarketplaceSchema, allow_none=True)
-    params = fields.Nested(ParamSchema, many=True)
-    tiers = fields.Nested(TierAccountsSchema)
-    items = fields.Nested(ItemSchema, many=True)
+    params = fields.Nested(ParamSchema, many=True, allow_none=True)
+    tiers = fields.Nested(TierAccountsSchema, allow_none=True)
+    items = fields.Nested(ItemSchema, many=True, allow_none=True)
     configuration = fields.Nested(ConfigurationSchema, allow_none=True)
 
     @post_load
@@ -566,17 +566,17 @@ class AssigneeField(fields.Field):
 
 
 class FulfillmentSchema(BaseSchema):
-    type = fields.Str()
-    created = fields.DateTime()
-    updated = fields.DateTime()
-    status = fields.Str()
-    params_form_url = fields.Str()
-    activation_key = fields.Str()
-    reason = fields.Str()
-    note = fields.Str()
-    asset = fields.Nested(AssetSchema)
-    contract = fields.Nested(ContractSchema, only=('id', 'name'))
-    marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'))
+    type = fields.Str(allow_none=True)
+    created = fields.DateTime(allow_none=True)
+    updated = fields.DateTime(allow_none=True)
+    status = fields.Str(allow_none=True)
+    params_form_url = fields.Str(allow_none=True)
+    activation_key = fields.Str(allow_none=True)
+    reason = fields.Str(allow_none=True)
+    note = fields.Str(allow_none=True)
+    asset = fields.Nested(AssetSchema, allow_none=True)
+    contract = fields.Nested(ContractSchema, only=('id', 'name'), allow_none=True)
+    marketplace = fields.Nested(MarketplaceSchema, only=('id', 'name'), allow_none=True)
     assignee = AssigneeField(allow_none=True)
 
     @post_load
@@ -586,16 +586,16 @@ class FulfillmentSchema(BaseSchema):
 
 
 class TierConfigSchema(BaseSchema):
-    name = fields.Str()
-    account = fields.Nested(TierAccountSchema)
-    product = fields.Nested(ProductSchema)
-    tier_level = fields.Int()
-    params = fields.Nested(ParamSchema, many=True)
-    connection = fields.Nested(ConnectionSchema)
+    name = fields.Str(allow_none=True)
+    account = fields.Nested(TierAccountSchema, allow_none=True)
+    product = fields.Nested(ProductSchema, allow_none=True)
+    tier_level = fields.Int(allow_none=True)
+    params = fields.Nested(ParamSchema, many=True, allow_none=True)
+    connection = fields.Nested(ConnectionSchema, allow_none=True)
     open_request = fields.Nested(BaseSchema, allow_none=True)
-    template = fields.Nested(TemplateSchema)
-    contract = fields.Nested(ContractSchema)
-    marketplace = fields.Nested(MarketplaceSchema)
+    template = fields.Nested(TemplateSchema, allow_none=True)
+    contract = fields.Nested(ContractSchema, allow_none=True)
+    marketplace = fields.Nested(MarketplaceSchema, allow_none=True)
     configuration = fields.Nested(ConfigurationSchema, allow_none=True)
     events = fields.Nested(EventsSchema, allow_none=True)
     status = fields.Str(allow_none=True)
@@ -607,15 +607,15 @@ class TierConfigSchema(BaseSchema):
 
 
 class TierConfigRequestSchema(BaseSchema):
-    type = fields.Str()
-    status = fields.Str()
-    configuration = fields.Nested(TierConfigSchema)
-    parent_configuration = fields.Nested(TierConfigSchema)
-    account = fields.Nested(TierAccountSchema)
-    product = fields.Nested(ProductSchema)
-    tier_level = fields.Int()
-    params = fields.Nested(ParamSchema, many=True)
-    environment = fields.Str()
+    type = fields.Str(allow_none=True)
+    status = fields.Str(allow_none=True)
+    configuration = fields.Nested(TierConfigSchema, allow_none=True)
+    parent_configuration = fields.Nested(TierConfigSchema, allow_none=True)
+    account = fields.Nested(TierAccountSchema, allow_none=True)
+    product = fields.Nested(ProductSchema, allow_none=True)
+    tier_level = fields.Int(allow_none=True)
+    params = fields.Nested(ParamSchema, many=True, allow_none=True)
+    environment = fields.Str(allow_none=True)
     assignee = fields.Nested(UserSchema, allow_none=True)
     template = fields.Nested(TemplateSchema, allow_none=True)
     reason = fields.Str(allow_none=True)
@@ -633,8 +633,8 @@ class TierConfigRequestSchema(BaseSchema):
 
 
 class UsageRecordsSchema(BaseSchema):
-    valid = fields.Int()
-    invalid = fields.Int()
+    valid = fields.Int(allow_none=True)
+    invalid = fields.Int(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -643,24 +643,24 @@ class UsageRecordsSchema(BaseSchema):
 
 
 class UsageFileSchema(BaseSchema):
-    name = fields.Str()
-    description = fields.Str()
-    note = fields.Str()
-    status = fields.Str()
-    created_by = fields.Str()
-    created_at = fields.Str()
-    upload_file_uri = fields.Str()
-    processed_file_uri = fields.Str()
-    product = fields.Nested(ProductSchema)
-    contract = fields.Nested(ContractSchema)
-    marketplace = fields.Nested(MarketplaceSchema)
-    vendor = fields.Nested(CompanySchema)
-    provider = fields.Nested(CompanySchema)
-    acceptance_note = fields.Str()
-    rejection_note = fields.Str()
-    error_details = fields.Str()
-    records = fields.Nested(UsageRecordsSchema)
-    events = fields.Nested(EventsSchema)
+    name = fields.Str(allow_none=True)
+    description = fields.Str(allow_none=True)
+    note = fields.Str(allow_none=True)
+    status = fields.Str(allow_none=True)
+    created_by = fields.Str(allow_none=True)
+    created_at = fields.Str(allow_none=True)
+    upload_file_uri = fields.Str(allow_none=True)
+    processed_file_uri = fields.Str(allow_none=True)
+    product = fields.Nested(ProductSchema, allow_none=True)
+    contract = fields.Nested(ContractSchema, allow_none=True)
+    marketplace = fields.Nested(MarketplaceSchema, allow_none=True)
+    vendor = fields.Nested(CompanySchema, allow_none=True)
+    provider = fields.Nested(CompanySchema, allow_none=True)
+    acceptance_note = fields.Str(allow_none=True)
+    rejection_note = fields.Str(allow_none=True)
+    error_details = fields.Str(allow_none=True)
+    records = fields.Nested(UsageRecordsSchema, allow_none=True)
+    events = fields.Nested(EventsSchema, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -669,14 +669,14 @@ class UsageFileSchema(BaseSchema):
 
 
 class UsageListingSchema(BaseSchema):
-    status = fields.Str()
-    contract = fields.Nested(ContractSchema)
-    product = fields.Nested(ProductSchema)
-    created = fields.Str()
+    status = fields.Str(allow_none=True)
+    contract = fields.Nested(ContractSchema, allow_none=True)
+    product = fields.Nested(ProductSchema, allow_none=True)
+    created = fields.Str(allow_none=True)
 
     # Undocumented fields (they appear in PHP SDK)
-    vendor = fields.Nested(CompanySchema)
-    provider = fields.Nested(CompanySchema)
+    vendor = fields.Nested(CompanySchema, allow_none=True)
+    provider = fields.Nested(CompanySchema, allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -685,14 +685,14 @@ class UsageListingSchema(BaseSchema):
 
 
 class UsageRecordSchema(BaseSchema):
-    usage_record_id = fields.Str()
-    item_search_criteria = fields.Str()
-    item_search_value = fields.Str()
-    quantity = fields.Int()
-    start_time_utc = fields.Str()
-    end_time_utc = fields.Str()
-    asset_search_criteria = fields.Str()
-    asset_search_value = fields.Str()
+    usage_record_id = fields.Str(allow_none=True)
+    item_search_criteria = fields.Str(allow_none=True)
+    item_search_value = fields.Str(allow_none=True)
+    quantity = fields.Int(allow_none=True)
+    start_time_utc = fields.Str(allow_none=True)
+    end_time_utc = fields.Str(allow_none=True)
+    asset_search_criteria = fields.Str(allow_none=True)
+    asset_search_value = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -701,10 +701,10 @@ class UsageRecordSchema(BaseSchema):
 
 
 class ConversationMessageSchema(BaseSchema):
-    conversation = fields.Str()
-    created = fields.DateTime()
-    creator = fields.Nested(UserSchema)
-    text = fields.Str()
+    conversation = fields.Str(allow_none=True)
+    created = fields.DateTime(allow_none=True)
+    creator = fields.Nested(UserSchema, allow_none=True)
+    text = fields.Str(allow_none=True)
 
     @post_load
     def make_object(self, data):
@@ -713,11 +713,11 @@ class ConversationMessageSchema(BaseSchema):
 
 
 class ConversationSchema(BaseSchema):
-    instance_id = fields.Str()
-    created = fields.DateTime()
-    topic = fields.Str()
-    messages = fields.Nested(ConversationMessageSchema, many=True)
-    creator = fields.Nested(UserSchema)
+    instance_id = fields.Str(allow_none=True)
+    created = fields.DateTime(allow_none=True)
+    topic = fields.Str(allow_none=True)
+    messages = fields.Nested(ConversationMessageSchema, many=True, allow_none=True)
+    creator = fields.Nested(UserSchema, allow_none=True)
 
     @post_load
     def make_object(self, data):